Simplify r_sign method

+++ src/rmath.h
@@ -56,7 +56,7 @@ decimal r_floor(const decimal&);
 decimal r_exp(const decimal&,bool round=true);
 decimal r_log(const decimal&,bool round=true);
 decimal r_log10(const decimal&,bool round=true);
-decimal r_sign(const decimal&);
+int r_sign(const decimal&);
 void r_rand_seed(int);
 decimal r_rand(const decimal&,const decimal&);
 void r_rand_clear();
